pyproject.toml
[project]
name = "geometric-solver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Nir Goren", email = "nir@gorenor.com" }
]
requires-python = ">=3.11"
dependencies = [
    "gradio>=3.36.1",
    "hydra-core>=1.3.2",
    "loguru>=0.7.0",
    "matplotlib>=3.7.0",
    "networkx>=3.0",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "pulp>=3.2.2",
    "pybind11>=3.0.1",
    "pyyaml>=6.0.2",
    "scikit-image>=0.21.0",
    "scipy>=1.16.1",
    "shapely>=2.0.0",
    "torch>=2.5.0",
    "torchvision>=0.18.0",
    "tqdm>=4.67.1",
    "wandb>=0.21.1",
]
